Inject cov props under parent
Creation order and inline components require
that the coverage definition is put under
the object the javascript is defined in.

Changed scriptcollector to find opening
curly bracket for the parent node and
store this in the found scripts.
When injecting properties, these are
now being replaced at individual positions
and not at the beginning of the root qml
component.